+++ takeout_lib/lib/player/handler.dart
@@ -50,7 +50,7 @@ class TakeoutPlayerHandler extends BaseAudioHandler with QueueHandler {
   final OffsetCacheRepository offsetRepository;
   final MediaRepository mediaRepository;
 
-  final AudioPlayer _player = AudioPlayer(maxSkipsOnError: 1);
+  final AudioPlayer _player = AudioPlayer(); //AudioPlayer(maxSkipsOnError: 1);
   final PlayCallback onPlay;
   final PauseCallback onPause;
   final StoppedCallback onStop;
@@ -480,7 +480,11 @@ class TakeoutPlayerHandler extends BaseAudioHandler with QueueHandler {
     // setAudioSource triggers events so use the correct index and position even though
     // skipToQueueItem does the same thing next.
     // also ensure _spiff is correct since events are triggered
-    await _player.setAudioSources(sources,
+    // await _player.setAudioSources(sources,
+    //     initialIndex: index, initialPosition: position);
+    final source = ConcatenatingAudioSource(children: []);
+    await source.addAll(sources);
+    await _player.setAudioSource(source,
         initialIndex: index, initialPosition: position);
 
     await skipToQueueItem(index);